webapp/requirements.txt
Django==1.9.5
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
Werkzeug==0.11.5
itsdangerous==0.24
